Create visualizers package
