+++ web-ui/src/main/java/org/archcnl/domain/input/visualization/coloredmodel/ColoredTriplet.java
@@ -4,7 +4,7 @@ import org.archcnl.domain.common.conceptsandrelations.Relation;
 import org.archcnl.domain.common.conceptsandrelations.andtriplets.triplet.ObjectType;
 import org.archcnl.domain.common.conceptsandrelations.andtriplets.triplet.Triplet;
 import org.archcnl.domain.common.conceptsandrelations.andtriplets.triplet.Variable;
-import org.archcnl.domain.input.visualization.rules.Cardinality;
+import org.archcnl.domain.input.visualization.visualizers.rules.Cardinality;
 
 public class ColoredTriplet extends Triplet {
