+++ web-ui/src/main/java/org/archcnl/domain/input/visualization/visualizers/Visualizer.java
@@ -1,4 +1,4 @@
-package org.archcnl.domain.input.visualization;
+package org.archcnl.domain.input.visualization.visualizers;
 
 public interface Visualizer {
