+++ web-ui/src/main/java/org/archcnl/domain/input/visualization/visualizers/mappings/ConceptVariant.java
@@ -1,4 +1,4 @@
-package org.archcnl.domain.input.visualization;
+package org.archcnl.domain.input.visualization.visualizers.mappings;
 
 import java.util.ArrayList;
 import java.util.Arrays;
