+++ web-ui/src/main/java/org/archcnl/domain/input/visualization/visualizers/rules/Cardinality.java
@@ -1,4 +1,4 @@
-package org.archcnl.domain.input.visualization.rules;
+package org.archcnl.domain.input.visualization.visualizers.rules;
 
 import org.archcnl.domain.input.visualization.exceptions.MappingToUmlTranslationFailedException;
