+++ web-ui/src/main/java/org/archcnl/domain/input/visualization/visualizers/rules/DomainRangeRuleVisualizer.java
@@ -1,4 +1,4 @@
-package org.archcnl.domain.input.visualization.rules;
+package org.archcnl.domain.input.visualization.visualizers.rules;
 
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -10,16 +10,20 @@ import org.archcnl.domain.input.model.architecturerules.ArchitectureRule;
 import org.archcnl.domain.input.visualization.coloredmodel.ColorState;
 import org.archcnl.domain.input.visualization.exceptions.MappingToUmlTranslationFailedException;
 
-public class UniversalRuleVisualizer extends RuleVisualizer {
+public class DomainRangeRuleVisualizer extends RuleVisualizer {
 
     private static final Pattern CNL_PATTERN =
             Pattern.compile(
-                    "Every " + "(a |an )?" + SUBJECT_REGEX + " can-only " + PHRASES_REGEX + "\\.");
+                    "Only " + "(a |an )?" + SUBJECT_REGEX + " can " + PHRASES_REGEX + "\\.");
 
-    public UniversalRuleVisualizer(
+    public DomainRangeRuleVisualizer(
             ArchitectureRule rule, ConceptManager conceptManager, RelationManager relationManager)
             throws MappingToUmlTranslationFailedException {
         super(rule, conceptManager, relationManager);
+    }
+
+    public static boolean matches(ArchitectureRule rule) {
+        return CNL_PATTERN.matcher(rule.toString()).matches();
     }
 
     @Override
@@ -33,12 +37,12 @@ public class UniversalRuleVisualizer extends RuleVisualizer {
         }
 
         RuleVariant wrong = new RuleVariant();
-        wrong.setSubjectTriplets(addPostfixToAllVariables(subjectTriplets, "W"));
+        var triplet =
+                RuleHelper.getTripletWithBaseType(
+                        subjectTriplets.get(0), conceptManager, usedVariables);
+        wrong.setSubjectTriplets(addPostfixToAllVariables(Arrays.asList(triplet), "W"));
         for (VerbPhrase phrase : verbPhrases.getPhrases()) {
-            var triplet = phrase.getObjectTriplets().get(0);
-            var coloredTriplet =
-                    RuleHelper.getTripletWithBaseType(triplet, conceptManager, usedVariables);
-            wrong.addObjectTriplets(addPostfixToAllVariables(Arrays.asList(coloredTriplet), "W"));
+            wrong.addObjectTriplets(addPostfixToAllVariables(phrase.getObjectTriplets(), "W"));
             wrong.addCopyOfPredicate(phrase.getPredicate());
         }
 
@@ -64,14 +68,13 @@ public class UniversalRuleVisualizer extends RuleVisualizer {
             correct.addCopyOfPredicate(phrase.getPredicate());
 
             RuleVariant wrong = new RuleVariant();
+            var triplet =
+                    RuleHelper.getTripletWithBaseType(
+                            subjectTriplets.get(0), conceptManager, usedVariables);
             wrong.setSubjectTriplets(
-                    addPostfixToAllVariables(subjectTriplets, wrongPostfix.toString()));
-            var triplet = phrase.getObjectTriplets().get(0);
-            var coloredTriplet =
-                    RuleHelper.getTripletWithBaseType(triplet, conceptManager, usedVariables);
+                    addPostfixToAllVariables(Arrays.asList(triplet), wrongPostfix.toString()));
             wrong.addObjectTriplets(
-                    addPostfixToAllVariables(
-                            Arrays.asList(coloredTriplet), wrongPostfix.toString()));
+                    addPostfixToAllVariables(phrase.getObjectTriplets(), wrongPostfix.toString()));
             wrong.addCopyOfPredicate(phrase.getPredicate());
 
             correct.setPredicateToColorState(ColorState.CORRECT);
@@ -83,10 +86,6 @@ public class UniversalRuleVisualizer extends RuleVisualizer {
             wrongPostfix.append("W");
         }
         return variants;
-    }
-
-    public static boolean matches(ArchitectureRule rule) {
-        return CNL_PATTERN.matcher(rule.toString()).matches();
     }
 
     @Override
