+++ web-ui/src/main/java/org/archcnl/domain/input/visualization/visualizers/rules/RuleVisualizer.java
@@ -1,4 +1,4 @@
-package org.archcnl.domain.input.visualization.rules;
+package org.archcnl.domain.input.visualization.visualizers.rules;
 
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -19,7 +19,6 @@ import org.archcnl.domain.common.conceptsandrelations.andtriplets.triplet.Triple
 import org.archcnl.domain.common.conceptsandrelations.andtriplets.triplet.Variable;
 import org.archcnl.domain.input.model.architecturerules.ArchitectureRule;
 import org.archcnl.domain.input.visualization.MappingTranslator;
-import org.archcnl.domain.input.visualization.Visualizer;
 import org.archcnl.domain.input.visualization.coloredmodel.ColoredTriplet;
 import org.archcnl.domain.input.visualization.coloredmodel.ColoredVariant;
 import org.archcnl.domain.input.visualization.coloredmodel.MappingSetter;
@@ -28,7 +27,8 @@ import org.archcnl.domain.input.visualization.diagram.PlantUmlPart;
 import org.archcnl.domain.input.visualization.exceptions.MappingToUmlTranslationFailedException;
 import org.archcnl.domain.input.visualization.helpers.MappingFlattener;
 import org.archcnl.domain.input.visualization.helpers.NamePicker;
-import org.archcnl.domain.input.visualization.rules.VerbPhraseContainer.Connector;
+import org.archcnl.domain.input.visualization.visualizers.Visualizer;
+import org.archcnl.domain.input.visualization.visualizers.rules.VerbPhraseContainer.Connector;
 
 public abstract class RuleVisualizer implements Visualizer {
