+++ web-ui/src/main/java/org/archcnl/domain/input/visualization/visualizers/rules/VerbPhrase.java
@@ -1,4 +1,4 @@
-package org.archcnl.domain.input.visualization.rules;
+package org.archcnl.domain.input.visualization.visualizers.rules;
 
 import java.util.List;
 import org.archcnl.domain.common.conceptsandrelations.andtriplets.triplet.Triplet;
